+++ lib/src/utilities/logger.dart
@@ -31,6 +31,7 @@ void infoLog([dynamic log, dynamic additional = "", bool? logInReleaseMode]) {
 
 void errorLog([dynamic log, dynamic additional = ""]) {
   var setting = SkadiProvider.loggerSetting;
-  if (kDebugMode || setting.error)
+  if (kDebugMode || setting.error) {
     debugPrint("Skadi Error Log: $log $additional");
+  }
 }